backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
pymongo
python-dotenv      # load .env in dev / docker
gunicorn           # production WSGI server
